backend/requirements.txt

# Web framework & server
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
pydantic>=1.10.0
python-multipart>=0.0.6

# Audio & speech
SpeechRecognition>=3.8.1
pydub>=0.25.1

# Optional/type helpers
typing-extensions>=4.0.0

python-dotenv>=1.0.0
redis>=5.0.0
docker>=6.0.0
openai>=1.0.0
pytest>=7.0.0
flake8>=6.0.0